Renamed airtraffic_framework -> air_traffic_framework for Pascal casing
