+++ lib/user_agent/browsers/air_traffic_framework.rb
