+++ spec/browsers/air_traffic_framework_user_agent_spec.rb
